make extension with vite and crxjs
