+++ package.json
@@ -1,0 +1,16 @@
+{
+	"name": "trade-aggregator",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "tsc && vite build",
+		"preview": "vite preview"
+	},
+	"devDependencies": {
+		"@crxjs/vite-plugin": "^2.0.0-beta.23",
+		"typescript": "^5.2.2",
+		"vite": "^5.2.0"
+	}
+}
